flask_endpoint/requirements.txt
flask
openai
flask-cors
python-dotenv
requests
python-dotenv
gunicorn